+++ SimService/src/main/java/com/simactivation/portal/SimService/repository/SimRepository.java
@@ -1,7 +1,8 @@
 package com.simactivation.portal.SimService.repository;
 
-import com.simactivation.portal.SimService.dto.SimDTO;
 import com.simactivation.portal.SimService.entity.SimDetails;
+import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Pattern;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -10,7 +11,7 @@ import java.util.Optional;
 @Repository
 public interface SimRepository extends CrudRepository<SimDetails,Integer> {
 
-    Optional<SimDetails> findByServiceNumber(int phoneNumber);
+    Optional<SimDetails> findByServiceNumber(long phoneNumber);
 
-    Optional<SimDetails> findByServiceNumberAndSimNumber(int serviceNumber, int simNumber);
+    Optional<SimDetails> findByServiceNumberAndSimNumber(long serviceNumber, long simNumber);
 }
